% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huePal.R
\name{huePal}
\alias{huePal}
\title{Create (named) single-hue gradient palette}
\usage{
huePal(
  hue = NULL,
  names = NULL,
  n = NULL,
  minChroma = 40,
  maxChroma = 150,
  minLum = 10,
  maxLum = 98,
  power = 0.8
)
}
\arguments{
\item{hue}{numeric hue value, or NULL to return palette function}

\item{names}{names for palette, or NULL for unnamed palette of length \code{n}}

\item{n}{number of unique shades in palette}

\item{minChroma}{minimum Chroma for palette}

\item{maxChroma}{maximum Chroma for palette}

\item{minLum}{minimum luminance for palette}

\item{maxLum}{maximum luminance for palette}

\item{power}{power used by \code{colorspace::sequential_hcl()}}
}
\value{
vector of colours, possibly named, or a function
}
\description{
Function to create create HCL palette of \code{n} colours based on fixed \code{hue}.
Luminance monotonically increases, whilst chroma increases and then decreases.

If names provided: return named palette of same length as unique(names),
with \code{n} distinct colours (if n is left null, all colours are unique)

If the \code{hue} value is left as NULL, a function will be returned,
which can generate a palette when given a \code{hue} values and \code{n} and/or \code{names}
}
\details{
Uses \code{colorspace::sequential_hcl()} with a fixed hue to generate palettes.

The palette generated will be roughly centered around the midpoint of
the luminance range, and at approximately the maximum chroma.

Note that edges of generated palette are cut off so min and max luminances
are never returned.
This is because they are, by default, too dark/light to be distinguishable
across hues.
}
\examples{
huePal(hue = 120, names = letters[1:9]) \%>\%
   scales::show_col(borders = NA)
huePal(hue = 120, names = letters[1:9]) == huePal(hue = 120, n = 9)
}
