% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hueSet.R
\name{hueSet}
\alias{hueSet}
\title{Get a set of equally spaced hues values}
\usage{
hueSet(n = 10, start = 180, cycles = 2)
}
\arguments{
\item{n}{number of hues to return}

\item{start}{starting hue value, in degrees from 0 to 359}

\item{cycles}{number of cycles rotating around the}
}
\value{
vector
}
\description{
Rotates around the 360 degrees of hue on the HCL colour wheel,

Starts at \code{start} and rotates \code{cycles} times around the wheel to obtain \code{n}
colours at equal intervals
}
\examples{
hueSet()

hueSet(cycles = 1)

hueSet(start = 0, cycles = 1)

hueSet(start = 0, n = 9, cycles = 3)
}
