% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeAfterN}
\alias{mergeAfterN}
\title{Merge values of vector after first n unique values}
\usage{
mergeAfterN(x, n, other = "other")
}
\arguments{
\item{x}{vector}

\item{n}{number of unique values/levels to keep}

\item{other}{name of new value/level to replace excess values with}
}
\value{
vector
}
\description{
Merge values of vector after first n unique values
}
\examples{
library(dplyr)
letters \%>\% mergeAfterN(15)
LETTERS \%>\% mergeAfterN(10, other = "?")

## Real data example ##
# works for factors too
gapminder::gapminder \%>\%
  dplyr::filter(year < 1970) \%>\%
  dplyr::pull(country) \%>\%
  mergeAfterN(10) \%>\%
  head(50)
}
