% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep_last}
\alias{rep_last}
\title{Repeat last value in vector to create longer vector}
\usage{
rep_last(x, length.out)
}
\arguments{
\item{x}{vector}

\item{length.out}{desired length of vector}
}
\description{
Helper function
}
\examples{
rep_last(letters[1:10], length.out = 15)
}
